Add community interaction helpers
